Improve code architecture for Docs Recipes
