+++ docs/content/recipes/cell-types/feedback-react/react/example1.jsx
