+++ docs/content/recipes/cell-types/flatpickr/javascript/example1.ts
@@ -2,277 +2,18 @@ import Handsontable from "handsontable/base";
 import { registerAllModules } from "handsontable/registry";
 import "handsontable/styles/handsontable.css";
 import "handsontable/styles/ht-theme-main.css";
-import multipleSelect from "multiple-select-vanilla";
-import type { MultipleSelectInstance } from "multiple-select-vanilla";
-import type { CellProperties } from "handsontable/settings";
+import { format, isDate } from "date-fns";
+import flatpickr from "flatpickr";
+import { CellProperties } from "handsontable/settings";
 import { editorFactory } from "handsontable/editors";
 import { rendererFactory } from "handsontable/renderers";
 // Register all Handsontable's modules.
 registerAllModules();
 
+const DATE_FORMAT_US = "MM/dd/yyyy";
+const DATE_FORMAT_EU = "dd/MM/yyyy";
+
 /* start:skip-in-preview */
-
-const components = [
-  { value: "1", label: "Component 1" },
-  { value: "2", label: "Component 2" },
-  { value: "3", label: "Component 3" },
-];
-
-const countryListAlpha2 = {
-  AF: "Afghanistan",
-  AL: "Albania",
-  DZ: "Algeria",
-  AS: "American Samoa",
-  AD: "Andorra",
-  AO: "Angola",
-  AI: "Anguilla",
-  AQ: "Antarctica",
-  AG: "Antigua and Barbuda",
-  AR: "Argentina",
-  AM: "Armenia",
-  AW: "Aruba",
-  AU: "Australia",
-  AT: "Austria",
-  AZ: "Azerbaijan",
-  BS: "Bahamas (the)",
-  BH: "Bahrain",
-  BD: "Bangladesh",
-  BB: "Barbados",
-  BY: "Belarus",
-  BE: "Belgium",
-  BZ: "Belize",
-  BJ: "Benin",
-  BM: "Bermuda",
-  BT: "Bhutan",
-  BO: "Bolivia (Plurinational State of)",
-  BQ: "Bonaire, Sint Eustatius and Saba",
-  BA: "Bosnia and Herzegovina",
-  BW: "Botswana",
-  BV: "Bouvet Island",
-  BR: "Brazil",
-  IO: "British Indian Ocean Territory (the)",
-  BN: "Brunei Darussalam",
-  BG: "Bulgaria",
-  BF: "Burkina Faso",
-  BI: "Burundi",
-  CV: "Cabo Verde",
-  KH: "Cambodia",
-  CM: "Cameroon",
-  CA: "Canada",
-  KY: "Cayman Islands (the)",
-  CF: "Central African Republic (the)",
-  TD: "Chad",
-  CL: "Chile",
-  CN: "China",
-  CX: "Christmas Island",
-  CC: "Cocos (Keeling) Islands (the)",
-  CO: "Colombia",
-  KM: "Comoros (the)",
-  CD: "Congo (the Democratic Republic of the)",
-  CG: "Congo (the)",
-  CK: "Cook Islands (the)",
-  CR: "Costa Rica",
-  HR: "Croatia",
-  CU: "Cuba",
-  CW: "Curaçao",
-  CY: "Cyprus",
-  CZ: "Czechia",
-  CI: "Côte d'Ivoire",
-  DK: "Denmark",
-  DJ: "Djibouti",
-  DM: "Dominica",
-  DO: "Dominican Republic (the)",
-  EC: "Ecuador",
-  EG: "Egypt",
-  SV: "El Salvador",
-  GQ: "Equatorial Guinea",
-  ER: "Eritrea",
-  EE: "Estonia",
-  SZ: "Eswatini",
-  ET: "Ethiopia",
-  FK: "Falkland Islands (the) [Malvinas]",
-  FO: "Faroe Islands (the)",
-  FJ: "Fiji",
-  FI: "Finland",
-  FR: "France",
-  GF: "French Guiana",
-  PF: "French Polynesia",
-  TF: "French Southern Territories (the)",
-  GA: "Gabon",
-  GM: "Gambia (the)",
-  GE: "Georgia",
-  DE: "Germany",
-  GH: "Ghana",
-  GI: "Gibraltar",
-  GR: "Greece",
-  GL: "Greenland",
-  GD: "Grenada",
-  GP: "Guadeloupe",
-  GU: "Guam",
-  GT: "Guatemala",
-  GG: "Guernsey",
-  GN: "Guinea",
-  GW: "Guinea-Bissau",
-  GY: "Guyana",
-  HT: "Haiti",
-  HM: "Heard Island and McDonald Islands",
-  VA: "Holy See (the)",
-  HN: "Honduras",
-  HK: "Hong Kong",
-  HU: "Hungary",
-  IS: "Iceland",
-  IN: "India",
-  ID: "Indonesia",
-  IR: "Iran (Islamic Republic of)",
-  IQ: "Iraq",
-  IE: "Ireland",
-  IM: "Isle of Man",
-  IL: "Israel",
-  IT: "Italy",
-  JM: "Jamaica",
-  JP: "Japan",
-  JE: "Jersey",
-  JO: "Jordan",
-  KZ: "Kazakhstan",
-  KE: "Kenya",
-  KI: "Kiribati",
-  KP: "Korea (the Democratic People's Republic of)",
-  KR: "Korea (the Republic of)",
-  KW: "Kuwait",
-  KG: "Kyrgyzstan",
-  LA: "Lao People's Democratic Republic (the)",
-  LV: "Latvia",
-  LB: "Lebanon",
-  LS: "Lesotho",
-  LR: "Liberia",
-  LY: "Libya",
-  LI: "Liechtenstein",
-  LT: "Lithuania",
-  LU: "Luxembourg",
-  MO: "Macao",
-  MG: "Madagascar",
-  MW: "Malawi",
-  MY: "Malaysia",
-  MV: "Maldives",
-  ML: "Mali",
-  MT: "Malta",
-  MH: "Marshall Islands (the)",
-  MQ: "Martinique",
-  MR: "Mauritania",
-  MU: "Mauritius",
-  YT: "Mayotte",
-  MX: "Mexico",
-  FM: "Micronesia (Federated States of)",
-  MD: "Moldova (the Republic of)",
-  MC: "Monaco",
-  MN: "Mongolia",
-  ME: "Montenegro",
-  MS: "Montserrat",
-  MA: "Morocco",
-  MZ: "Mozambique",
-  MM: "Myanmar",
-  NA: "Namibia",
-  NR: "Nauru",
-  NP: "Nepal",
-  NL: "Netherlands (the)",
-  NC: "New Caledonia",
-  NZ: "New Zealand",
-  NI: "Nicaragua",
-  NE: "Niger (the)",
-  NG: "Nigeria",
-  NU: "Niue",
-  NF: "Norfolk Island",
-  MP: "Northern Mariana Islands (the)",
-  NO: "Norway",
-  OM: "Oman",
-  PK: "Pakistan",
-  PW: "Palau",
-  PS: "Palestine, State of",
-  PA: "Panama",
-  PG: "Papua New Guinea",
-  PY: "Paraguay",
-  PE: "Peru",
-  PH: "Philippines (the)",
-  PN: "Pitcairn",
-  PL: "Poland",
-  PT: "Portugal",
-  PR: "Puerto Rico",
-  QA: "Qatar",
-  MK: "Republic of North Macedonia",
-  RO: "Romania",
-  RU: "Russian Federation (the)",
-  RW: "Rwanda",
-  RE: "Réunion",
-  BL: "Saint Barthélemy",
-  SH: "Saint Helena, Ascension and Tristan da Cunha",
-  KN: "Saint Kitts and Nevis",
-  LC: "Saint Lucia",
-  MF: "Saint Martin (French part)",
-  PM: "Saint Pierre and Miquelon",
-  VC: "Saint Vincent and the Grenadines",
-  WS: "Samoa",
-  SM: "San Marino",
-  ST: "Sao Tome and Principe",
-  SA: "Saudi Arabia",
-  SN: "Senegal",
-  RS: "Serbia",
-  SC: "Seychelles",
-  SL: "Sierra Leone",
-  SG: "Singapore",
-  SX: "Sint Maarten (Dutch part)",
-  SK: "Slovakia",
-  SI: "Slovenia",
-  SB: "Solomon Islands",
-  SO: "Somalia",
-  ZA: "South Africa",
-  GS: "South Georgia and the South Sandwich Islands",
-  SS: "South Sudan",
-  ES: "Spain",
-  LK: "Sri Lanka",
-  SD: "Sudan (the)",
-  SR: "Suriname",
-  SJ: "Svalbard and Jan Mayen",
-  SE: "Sweden",
-  CH: "Switzerland",
-  SY: "Syrian Arab Republic",
-  TW: "Taiwan",
-  TJ: "Tajikistan",
-  TZ: "Tanzania, United Republic of",
-  TH: "Thailand",
-  TL: "Timor-Leste",
-  TG: "Togo",
-  TK: "Tokelau",
-  TO: "Tonga",
-  TT: "Trinidad and Tobago",
-  TN: "Tunisia",
-  TR: "Turkey",
-  TM: "Turkmenistan",
-  TC: "Turks and Caicos Islands (the)",
-  TV: "Tuvalu",
-  UG: "Uganda",
-  UA: "Ukraine",
-  AE: "United Arab Emirates (the)",
-  GB: "United Kingdom of Great Britain and Northern Ireland (the)",
-  UM: "United States Minor Outlying Islands (the)",
-  US: "United States of America (the)",
-  UY: "Uruguay",
-  UZ: "Uzbekistan",
-  VU: "Vanuatu",
-  VE: "Venezuela (Bolivarian Republic of)",
-  VN: "Viet Nam",
-  VG: "Virgin Islands (British)",
-  VI: "Virgin Islands (U.S.)",
-  WF: "Wallis and Futuna",
-  EH: "Western Sahara",
-  YE: "Yemen",
-  ZM: "Zambia",
-  ZW: "Zimbabwe",
-  AX: "Åland Islands",
-};
-
-export const coutries = Object.entries(countryListAlpha2)
-  .map(([value, label]) => ({ value, label }))
-  .slice(0, 198); // 199 is the max number of options that can be displayed in the dropdown
 
 const inputData = [
   {
@@ -619,24 +360,6 @@ const inputData = [
 
 export const data = inputData.map((el) => ({
   ...el,
-  components: components
-    .map((n) => {
-      return [Math.random(), n];
-    })
-    .sort()
-    .map((n) => {
-      return n[1];
-    })
-    .slice(0, Math.ceil(Math.random() * components.length)),
-  countries: coutries
-    .map((n) => {
-      return [Math.random(), n];
-    })
-    .sort()
-    .map((n) => {
-      return n[1];
-    })
-    .slice(0, Math.ceil(Math.random() * 5)),
 }));
 
 /* end:skip-in-preview */
@@ -648,54 +371,54 @@ const cellDefinition: Pick<
   CellProperties,
   "renderer" | "validator" | "editor"
 > = {
-  renderer: rendererFactory(({ td, value }) => {
-    td.innerHTML = value.length > 0
-      ? value.map((el: { label: string }) => el.label).join(", ")
-      : "No elements";
-
-    return td;
+  validator: (value, callback) => {
+    callback(isDate(new Date(value)));
+  },
+  renderer: rendererFactory(({ td, value, cellProperties }) => {
+    td.innerText = format(new Date(value), cellProperties.renderFormat);
   }),
-  editor: editorFactory<
-    { input: HTMLSelectElement; multiselect: MultipleSelectInstance }
-  >({
+  editor: editorFactory<{
+    input: HTMLInputElement;
+    flatpickr: flatpickr.Instance;
+    flatpickrSettings: flatpickr.Options.Options;
+  }>({
     init(editor) {
+      // create the input element on init. This is a text input that color picker will be attached to.
       editor.input = editor.hot.rootDocument.createElement(
-        "SELECT",
-      ) as HTMLSelectElement;
-      editor.input.setAttribute("multiple", "multiple");
-      editor.input.setAttribute("data-multi-select", "");
-      editor.multiselect = multipleSelect(
-        editor.input,
-      ) as MultipleSelectInstance;
+        "INPUT",
+      ) as HTMLInputElement;
+      editor.flatpickr = flatpickr(editor.input, {
+        dateFormat: "Y-m-d",
+        enableTime: false,
+        onChange: () => {
+          editor.finishEditing();
+        },
+      });
+      /**
+       * Prevent recognizing clicking on datepicker as clicking outside of table.
+       */
+      editor.hot.rootDocument.addEventListener("mousedown", (event) => {
+        if (editor.flatpickr.calendarContainer.contains(event.target as Node)) {
+          event.stopPropagation();
+        }
+      });
     },
+    beforeOpen(editor, { originalValue, cellProperties }) {
+      editor.setValue(originalValue);
 
-    beforeOpen(editor, { cellProperties }) {
-      editor.input.innerHTML = cellProperties?.selectMultipleOptions
-        ?.map((el: { value: string; label: string }) =>
-          `<option value="${el.value}">${el.label}</option>`
-        )
-        .join("");
-      editor.multiselect.refresh();
-    },
-    afterOpen(editor) {
-      editor.multiselect.open();
+      for (const key in cellProperties.flatpickrSettings) {
+        editor.flatpickr.set(
+          key as keyof flatpickr.Options.Options,
+          cellProperties.flatpickrSettings[key],
+        );
+      }
     },
     getValue(editor) {
-      return Array.from(editor.input.options)
-        .filter((option) => option.selected)
-        .map((option) => ({ value: option.value, label: option.label }));
+      return editor.input.value;
     },
     setValue(editor, value) {
-      // https://github.com/handsontable/handsontable/issues/3510
-      value = typeof value === "string" ? editor.originalValue : value;
-      Array.from(editor.input.options).forEach(
-        (
-          option,
-        ) => (option.selected = value.some((el: { value: string }) =>
-          el.value === option.value
-        )),
-      );
-      editor.multiselect.refresh();
+      editor.input.value = value;
+      editor.flatpickr.setDate(new Date(value));
     },
   }),
 };
@@ -704,32 +427,43 @@ const cellDefinition: Pick<
 const hotOptions: Handsontable.GridSettings = {
   themeName: "ht-theme-main",
   data,
-  colHeaders: ["ID", "Item Name", "Components", "Countries"],
+  colHeaders: ["ID", "Item Name", "Restock Date UE", "Restock Date US"],
   autoRowSize: true,
   rowHeaders: true,
   height: "auto",
+  autoWrapRow: true,
   columns: [
-    { data: "id", type: "numeric" },
-    {
-      data: "countries",
-      width: 150,
-      allowInvalid: false,
-      ...cellDefinition,
-      selectMultipleOptions: coutries,
-    },
+    { data: "id", type: "numeric", width: 150 },
     {
       data: "itemName",
       type: "text",
+      width: 150,
     },
     {
-      data: "components",
+      data: "restockDate",
       width: 150,
       allowInvalid: false,
       ...cellDefinition,
-      selectMultipleOptions: components,
+      renderFormat: DATE_FORMAT_EU,
+      flatpickrSettings: {
+        locale: {
+          firstDayOfWeek: 1,
+        },
+      },
+    },
+    {
+      data: "restockDate",
+      width: 150,
+      allowInvalid: false,
+      ...cellDefinition,
+      renderFormat: DATE_FORMAT_US,
+      flatpickrSettings: {
+        locale: {
+          firstDayOfWeek: 0,
+        },
+      },
     },
   ],
-  autoWrapRow: true,
   licenseKey: "non-commercial-and-evaluation",
 };
