+++ docs/content/recipes/sidebar.js
@@ -1,17 +1,16 @@
-const cellsItems = [
-  { path: 'cells/guide-feedback-react/guide-feedback', title: 'Simple Feedback', onlyFor: ['react'] },
-  { path: 'cells/guide-feedback/guide-feedback', title: 'Simple Feedback', onlyFor: ['javascript'] },
-  { path: 'cells/guide-rating/guide-rating', title: 'Stars Rating', onlyFor: ['javascript'] },
-  { path: 'cells/guide-color-picker/guide-color-picker', title: 'Color picker', onlyFor: ['javascript'] },
-  { path: 'cells/guide-flatpickr/guide-flatpickr', title: 'Datetime `flatpickr` picker', onlyFor: ['javascript'] },
-  { path: 'cells/guide-select-multiple/guide-select-multiple', title: 'Multiple Select Dropdown', onlyFor: ['javascript'] },
-  { path: 'cells/guide-pikaday/guide-pikaday', title: 'Date picker pikaday', onlyFor: ['javascript'] },
+const cellTypesItems = [
+  { path: 'cell-types/color-picker/color-picker', title: 'Color picker', onlyFor: ['javascript'] },
+  { path: 'cell-types/feedback-react/feedback', title: 'Simple Feedback', onlyFor: ['react'] },
+  { path: 'cell-types/feedback/feedback', title: 'Simple Feedback', onlyFor: ['javascript'] },
+  { path: 'cell-types/flatpickr/flatpickr', title: 'Datetime `flatpickr` picker', onlyFor: ['javascript'] },
+  { path: 'cell-types/select-multiple/select-multiple', title: 'Multiple Select Dropdown', onlyFor: ['javascript'] },
+  { path: 'cell-types/pikaday/pikaday', title: 'Date picker pikaday', onlyFor: ['javascript'] },
+  { path: 'cell-types/rating/rating', title: 'Stars Rating', onlyFor: ['javascript'] },
 ];
 
 module.exports = {
   sidebar: [
-    //{ title: 'introduction', path: 'introduction', children: [] },
     'introduction',
-    { title: 'Cell Definitions', path: 'cells', children: cellsItems, collapsable: false, onlyFor: ['react', 'javascript'] },
+    { title: 'Cell Types', path: 'cell-types', children: cellTypesItems, collapsable: false, onlyFor: ['react', 'javascript'] },
   ],
 };
